feat: added lock/unlock intial functions

+++ DoorHandle/Gpt/Gpt.h
@@ -1,8 +1,8 @@
 /*
- * Gpt.h
+ * main.c
  *
  *  Created on: May 21, 2023
- *      Author: I1bra
+ *  Author: Ibrahim Mohamed
  */
 
 #ifndef GPT_H_
@@ -11,7 +11,7 @@
 #include "Std_Types.h"
 
 void GPT_Init(void);
-void GPT_StartTimer(unsigned long int OverFlowTicks);
+void GPT_StartTimer(uint32 int OverFlowTicks);
 unsigned char GPT_CheckTimeIsElapsed(void);
 unsigned long int GPT_GetElapsedTime(void);
 unsigned long int GPT_GetRemainingTime(void);
